pyproject.toml
[project]
name = "country-universities-llm"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "aiohttp>=3.10",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
  "tenacity>=9.0",
  "openai>=1.48.0",
]

[tool.poetry]
package-mode = false   

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
